Front End Task Completed
GUVI Team gave a task to develop a sign up page, login page and profile page using React JS. Here I had created those three pages successfully. And there is another constraint, which is user is not logged in then profile page should not be visible, redirect to login page for user login use the service. If the user is logged in then login page or sign up page should not visible because user is already registered.